Separated servlet tests
